pyproject.toml
[project]
name = "supervisedgenksc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
packages = [{ include = "src" }] 
requires-python = ">=3.11"
dependencies = [
    "h5py>=3.13.0",
    "imageio>=2.37.0",
    "lightning>=2.5.0.post0",
    "matplotlib>=3.10.0",
    "nested-dict>=1.61",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "setuptools>=75.8.0",
    "tensorboard>=2.19.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "typing>=3.10.0.0",
]

